+++ components/Bear.jsx
