+++ components/CardFull.jsx
