+++ components/HeroAnimate.jsx
