--- src/pages/_app.jsx
@@ -1,5 +1,0 @@
-import '@/styles/globals.css'
-
-export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />
-}
