--- src/pages/dev.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import Navbar from '../components/Navbar'
-
-const dev = () => {
-  return (
-    <div>
-    <Navbar />
-    </div>
-  )
-}
-
-export default dev
