--- src/pages/question1.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import Footer from '../components/Footer'
-import Header from '../components/Header'
-
-
-export default function Question1() {
-    return (
-        <div>
-            <Header name="User Name" />
-            Question1
-            <Footer />
-        </div>
-    )
-}
