--- src/pages/scoreboard.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import Footer from '../components/Footer'
-import Header from '../components/Header'
-
-export default function Scoreboard() {
-  return (
-    <div>
-      <Header name="User Name" />
-      Scoreboard
-      <Footer />
-    </div>
-  )
-}
